% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quick_save}
\alias{quick_save}
\alias{quick_load}
\title{Quick save of a single object}
\usage{
quick_save(object)

quick_load()
}
\arguments{
\item{object}{object to save}
}
\description{
Saves object to ~/.qstools directory, together with quick_load() allows
to quickly move object between R sessions (e.g. between the analysis and
package debuggin session)
}
\section{Functions}{
\itemize{
\item \code{quick_load()}: Quick load of ~/.qstools/object.Rds

}}
