% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabset.R
\name{tabset}
\alias{tabset}
\title{Print or plot list of elements as quarto tabset}
\usage{
tabset(lst, ..., heading = "##", fun = print)
}
\arguments{
\item{lst}{List of elements to print or plot}

\item{...}{not used}

\item{heading}{Heading to use for each tab}

\item{fun}{Plotting or printing function}
}
\description{
Remember to use this function inside a tabset div and quarto chunk with
results = "asis" option. See examples below
}
\examples{
\dontrun{
:::: {.column-page-outset}
::: {.panel-tabset}
```{r, results='asis'}
x <- list(z = 1:10, y = rnorm(100))
tabset(x)
```
:::
::::
}

}
