% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow.R
\name{write_nf_session_info}
\alias{write_nf_session_info}
\title{Write session info to a YAML file in a nf-core manner}
\usage{
write_nf_session_info(
  task_process,
  ...,
  pkgs = "other_packages",
  path = "versions.yml"
)
}
\arguments{
\item{task_process}{Name of the process}

\item{...}{Not used}

\item{pkgs}{One of:
\itemize{
\item "other_packages" - lists main loaded packages
\item "all" - lists all loaded packages
\item vector of packages to be listed
}}

\item{path}{Path to the output file}
}
\description{
Write session info to a YAML file in a nf-core manner
}
\examples{
library(qsmisc)
write_nf_session_info("PROCESSX":, path = "versions.yml")

\dontrun{
write_nf_session_info("${task.process}":, path = "versions.yml")
}
}
